Add capybara, factory_bot_rails, and clearance
Clearance is a library for helping to maintain user accounts, their passwords,
and resets. It's got a bunch of helpers and I have to use it for work sometimes.
So it's nice to have here. It lets you do a bunch of stuff and auto generate
almost all of these files.

I've also added some of the changes to add the syntax and what not to make
testing more reasonable. You can now do something like "visit root_path(as: user)" and have that just work

I've only added the sign-in links to the home page (our only page).

I also generated the default clearance specs for regression testing. I don't
have any email client lib setup but we might want to reach for mailgun or
something at some point as well