Use Rack::Deflator
There's an excellent older blog post about this from a guy named Richard
Schneeman about what this does and how it's useful for us.
https://schneems.com/2017/11/08/80-smaller-rails-footprint-with-rack-deflate/

This still works and is awesome because we a) can cache the same gzipped
response with the browser and reuse and b) cpus are stupid fast and network IO
is almost always the challenge with things keeping things small and fast.

On my windows box running in the WSL mode with browser cache (granted this is
just on the home page) it's about a 10% reduction in page load time served
locally. On a remote server this should only be faster.

I figure we leave this in, if it works out for us, we're golden. If not we can
always rip it out