Add Collection, Records, and the Join Table CollectionRecords
The idea here is that a user may want to store collections of records in an
accessible way. We can imagine this as maybe a way to track battle lists,
painting progress, or even plan out hobby spending. For now we are only
sketching out the general shape of these models and not getting into the full
functionality